chore: update bridge example schema
