+++ bridge/src/stackr/machine.ts
@@ -1,26 +1,8 @@
-import { State, StateMachine } from "@stackr/sdk/machine";
-import { solidityPackedKeccak256 } from "ethers";
+import { StateMachine } from "@stackr/sdk/machine";
 
 import * as genesisState from "../../genesis-state.json";
+import { Balances, BridgeState } from "./state";
 import { transitions } from "./transitions";
-
-type Balances = {
-  address: string;
-  balance: number;
-}[];
-
-export class BridgeState extends State<Balances> {
-  constructor(state: Balances) {
-    super(state);
-  }
-
-  getRootHash(): string {
-    // the following line is for testing purposes only
-    // in production, the root hash should be calculated
-    // by creating a merkle tree of the state leaves
-    return solidityPackedKeccak256(["string"], [JSON.stringify(this.state)]);
-  }
-}
 
 const machine = new StateMachine({
   id: "bridge",
