+++ bridge/src/stackr/schemas.ts
