+++ bridge/src/utils.ts
@@ -1,0 +1,15 @@
+import { ActionSchema, AllowedInputTypes } from "@stackr/sdk";
+import { Wallet } from "ethers";
+
+export const signMessage = async (
+  wallet: Wallet,
+  schema: ActionSchema,
+  payload: AllowedInputTypes
+) => {
+  const signature = await wallet.signTypedData(
+    schema.domain,
+    schema.EIP712TypedData.types,
+    payload
+  );
+  return signature;
+};
